train/requirements.txt
SQLAlchemy==2.0.31
psycopg2-binary==2.9.10
boto3==1.39.14
python-dotenv==1.0.1
PyPDF2>=3.0.0
transformers==4.54.1
datasets>=2.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
accelerate>=0.26.0
mlflow==3.1.2